+++ lib/handy.dart
@@ -5,3 +5,4 @@ export 'src/range.dart';
 export 'src/random.dart';
 export 'src/cleaner.dart';
 export 'src/datetime.dart';
+export 'src/cache.dart';